pybooker/requirements.txt
djangorestframework==3.14.0
Django==4.1.3
requests==2.28.1
Pillow==9.3.0
django-debug-toolbar==3.7.0
psycopg2-binary==2.9.5
django-environ==0.9.0
python-dotenv==0.21.0
